aider: Fix type mismatch in bbContains function by converting float64 values to float32 before comparison.
# Aider chat conversation:

USER: I ran this command:

 go vet ./...

And got this output:

# github.com/chooban/progdl-go/internal/pdf
internal/pdf/pdf.go:198:24: invalid operation: textBb.Left >= objBb.Left (mismatched types float64 and float32)
internal/pdf/pdf.go:199:19: invalid operation: textBb.Right <= objBb.Right (mismatched types float64 and float32)
internal/pdf/pdf.go:200:17: invalid operation: textBb.Top >= objBb.Top (mismatched types float64 and float32)
internal/pdf/pdf.go:201:20: invalid operation: textBb.Bottom <= objBb.Bottom (mismatched types float64 and float32)

ASSISTANT: The error message indicates that there is a type mismatch in the comparison operations. The `textBb` fields are of type `float64` while the `objBb` fields are of type `float32`.

To fix this, we need to convert the `float64` values to `float32` before the comparison. Here is the *SEARCH/REPLACE block* for the change:

```go
internal/pdf/pdf.go
<<<<<<< SEARCH
// bbContains returns true if the textBb is contained completely within objBb
func bbContains(objBb responses.FPDFPageObj_GetBounds, textBb responses.FPDFText_GetRect) bool {
	return textBb.Left >= objBb.Left &&
		textBb.Right <= objBb.Right &&
		textBb.Top >= objBb.Top &&
		textBb.Bottom <= objBb.Bottom
}
=======
// bbContains returns true if the textBb is contained completely within objBb
func bbContains(objBb responses.FPDFPageObj_GetBounds, textBb responses.FPDFText_GetRect) bool {
	return float32(textBb.Left) >= objBb.Left &&
		float32(textBb.Right) <= objBb.Right &&
		float32(textBb.Top) >= objBb.Top &&
		float32(textBb.Bottom) <= objBb.Bottom
}
>>>>>>> REPLACE
```

This change converts the `float64` values to `float32` before comparing them with the `objBb` fields. This should resolve the type mismatch error.

+++ internal/pdf/pdf.go
@@ -195,8 +195,8 @@ func (p *PdfiumReader) Credits(filename string, pageNumber int) (contents string
 
 // bbContains returns true if the textBb is contained completely within objBb
 func bbContains(objBb responses.FPDFPageObj_GetBounds, textBb responses.FPDFText_GetRect) bool {
-	return textBb.Left >= objBb.Left &&
-		textBb.Right <= objBb.Right &&
-		textBb.Top >= objBb.Top &&
-		textBb.Bottom <= objBb.Bottom
+	return float32(textBb.Left) >= objBb.Left &&
+		float32(textBb.Right) <= objBb.Right &&
+		float32(textBb.Top) >= objBb.Top &&
+		float32(textBb.Bottom) <= objBb.Bottom
 }
